refactored tsconfig and vit config paths/aliases added js docs for components and index.ts files
