+++ src/main.tsx
@@ -1,8 +1,10 @@
+// react dom
 import ReactDOM from 'react-dom/client';
+// react router
 import { RouterProvider } from 'react-router-dom';
-import ThemeProvider from 'src/theme/Theme';
-
-import { router } from 'src/router';
+// components
+import { router } from './router';
+import { ThemeProvider } from './theme';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <ThemeProvider>
