+++ src/pages/Error/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Error';
+export * from './Error';
