+++ src/pages/Home/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Home';
+export * from './Home';
