+++ src/pages/Roles/Roles.tsx
@@ -1,7 +1,5 @@
 interface IRolesProps {}
 
-const Roles: React.FC<IRolesProps> = (props) => {
+export const Roles: React.FC<IRolesProps> = (props) => {
   return <div>Roles Page</div>;
 };
-
-export default Roles;
