+++ src/pages/Roles/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Roles';
+export * from './Roles';
