+++ src/pages/Users/Users.tsx
@@ -1,7 +1,5 @@
 interface IUserProps {}
 
-const Users: React.FC<IUserProps> = (props) => {
+export const Users: React.FC<IUserProps> = (props) => {
   return <div>Users Page</div>;
 };
-
-export default Users;
