+++ src/pages/Users/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Users';
+export * from './Users';
