+++ src/router/Routes.tsx
@@ -1,10 +1,13 @@
+// react router
 import { createBrowserRouter } from 'react-router-dom';
 
-import Layout from 'src/theme/layout/Layout';
-import Home from 'src/pages/Home/Home';
-import Users from 'src/pages/Users/Users';
-import Roles from 'src/pages/Roles/Roles';
-import Error from 'src/pages/Error/Error';
+// ----------------------------------------------------------------
+
+import { Layout } from '@theme/layout';
+import { Home } from '@pages/Home';
+import { Users } from '@pages/Users';
+import { Roles } from '@pages/Roles';
+import { Error } from '@pages/Error';
 
 export const router = createBrowserRouter([
   {
