--- src/theme/Theme.tsx
@@ -1,22 +1,0 @@
-import {
-  createTheme,
-  ThemeProvider as MuiThemeProvide,
-  CssBaseline
-} from '@mui/material';
-
-interface IThemeProvider {
-  children: React.ReactNode;
-}
-
-const theme = createTheme({});
-
-const ThemeProvider: React.FC<IThemeProvider> = ({ children }) => {
-  return (
-    <MuiThemeProvide theme={theme}>
-      <CssBaseline />
-      {children}
-    </MuiThemeProvide>
-  );
-};
-
-export default ThemeProvider;
