+++ src/theme/layout/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Layout';
+export * from './Layout';
