+++ tsconfig.json
@@ -21,12 +21,13 @@
     "noFallthroughCasesInSwitch": true,
     "baseUrl": ".",
     "paths": {
-      // "@src/*": ["src/*"],
-      "@components/*": ["src/components/*"],
-      "@libs/*": ["src/libs/*"],
-      "@pages/*": ["src/pages/*"],
-      "@router/*": ["src/router/*"],
-      "@config/*": ["src/config/*"]
+      "@/*": ["./src/*"],
+      "@components/*": ["./src/components/*"],
+      "@libs/*": ["./src/libs/*"],
+      "@pages/*": ["./src/pages/*"],
+      "@router/*": ["./src/router/*"],
+      "@config/*": ["./src/config/*"],
+      "@theme/*": ["./src/theme/*"]
     }
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"],
